+++ click_extended/transform/as_path.py
@@ -10,14 +10,28 @@ import os
 from pathlib import Path
 
 from click_extended.core._child_node import ChildNode, ProcessContext
-from click_extended.errors import ValidationError
+from click_extended.errors import UnhandledValueError, ValidationError
 from click_extended.types import Decorator
+from click_extended.utils import (
+    is_nested_tuple_value,
+    is_single_value,
+    is_tuple_value,
+)
 
 
 class AsPath(ChildNode):
     """Transformer decorator to convert a string to a Path."""
 
-    def process(self, value: str | Path, context: ProcessContext) -> Path:
+    def process(
+        self,
+        value: (
+            str
+            | Path
+            | tuple[str | Path, ...]
+            | tuple[tuple[str | Path, ...], ...]
+        ),
+        context: ProcessContext,
+    ) -> Path | tuple[Path, ...] | tuple[tuple[Path, ...], ...]:
         exists = context.kwargs["exists"]
         parents = context.kwargs["parents"]
         extensions = context.kwargs["extensions"]
@@ -33,91 +47,103 @@ class AsPath(ChildNode):
         is_writable = context.kwargs["is_writable"]
         is_executable = context.kwargs["is_executable"]
 
-        path = Path(value) if isinstance(value, str) else value
-        path = path.expanduser()
-        path = path.resolve() if follow_symlink else path.absolute()
+        def validate_path(v: str | Path) -> Path:
+            """Validate and transform a single path value."""
+            path = Path(v) if isinstance(v, str) else v
+            path = path.expanduser()
+            path = path.resolve() if follow_symlink else path.absolute()
 
-        # Symlinks
-        if not allow_symlink and path.is_symlink():
-            raise ValidationError(
-                f"Path '{path}' is a symlink, but symlinks are not allowed."
-            )
-
-        # Existence
-        if exists and not path.exists():
-            raise ValidationError(f"Path '{path}' does not exist.")
-
-        # Parents
-        if parents and not path.parent.exists():
-            raise ValidationError(
-                f"Parent directory '{path.parent}' does not exist."
-            )
-
-        if path.exists():
-            is_file = path.is_file()
-            is_dir = path.is_dir()
-
-            # File disallowed
-            if is_file and not allow_file:
+            # Symlinks
+            if not allow_symlink and path.is_symlink():
                 raise ValidationError(
-                    f"Path '{path}' is a file, but files are not allowed."
+                    f"Path '{path}' is a symlink, but symlinks are not allowed."
                 )
 
-            # Directory disallowed
-            if is_dir and not allow_directory:
+            # Existence
+            if exists and not path.exists():
+                raise ValidationError(f"Path '{path}' does not exist.")
+
+            # Parents
+            if parents and not path.parent.exists():
                 raise ValidationError(
-                    f"Path '{path}' is a directory, "
-                    "but directories are not allowed."
+                    f"Parent directory '{path.parent}' does not exist."
                 )
 
-            # Check empty directory
-            if is_dir and not allow_empty_directory:
-                if not any(path.iterdir()):
+            if path.exists():
+                is_file = path.is_file()
+                is_dir = path.is_dir()
+
+                # File disallowed
+                if is_file and not allow_file:
                     raise ValidationError(
-                        f"Directory '{path}' is empty, but empty directories "
-                        f"are not allowed."
+                        f"Path '{path}' is a file, but files are not allowed."
                     )
 
-            # Check empty file
-            if is_file and not allow_empty_file:
-                if path.stat().st_size == 0:
+                # Directory disallowed
+                if is_dir and not allow_directory:
                     raise ValidationError(
-                        f"File '{path}' is empty, but empty files are not "
-                        f"allowed."
+                        f"Path '{path}' is a directory, "
+                        "but directories are not allowed."
                     )
 
-            # Check permissions
-            if is_readable and not os.access(path, os.R_OK):
-                raise ValidationError(f"Path '{path}' is not readable.")
+                # Check empty directory
+                if is_dir and not allow_empty_directory:
+                    if not any(path.iterdir()):
+                        raise ValidationError(
+                            f"Directory '{path}' is empty, but empty "
+                            f"directories are not allowed."
+                        )
 
-            if is_writable and not os.access(path, os.W_OK):
-                raise ValidationError(f"Path '{path}' is not writable.")
+                # Check empty file
+                if is_file and not allow_empty_file:
+                    if path.stat().st_size == 0:
+                        raise ValidationError(
+                            f"File '{path}' is empty, but empty files are not "
+                            f"allowed."
+                        )
 
-            if is_executable and not os.access(path, os.X_OK):
-                raise ValidationError(f"Path '{path}' is not executable.")
+                # Check permissions
+                if is_readable and not os.access(path, os.R_OK):
+                    raise ValidationError(f"Path '{path}' is not readable.")
 
-        # Check extensions
-        if extensions and (not path.exists() or path.is_file()):
-            if not any(path.name.endswith(ext) for ext in extensions):
+                if is_writable and not os.access(path, os.W_OK):
+                    raise ValidationError(f"Path '{path}' is not writable.")
+
+                if is_executable and not os.access(path, os.X_OK):
+                    raise ValidationError(f"Path '{path}' is not executable.")
+
+            # Check extensions
+            if extensions and (not path.exists() or path.is_file()):
+                if not any(path.name.endswith(ext) for ext in extensions):
+                    raise ValidationError(
+                        f"Path '{path}' does not have an allowed extension. "
+                        f"Allowed extensions: {', '.join(extensions)}"
+                    )
+
+            # Patterns
+            if include_pattern:
+                if not fnmatch.fnmatch(path.name, include_pattern):
+                    raise ValidationError(
+                        f"Path '{path}' does not match include pattern "
+                        f"'{include_pattern}'."
+                    )
+            elif exclude_pattern and fnmatch.fnmatch(
+                path.name, exclude_pattern
+            ):
                 raise ValidationError(
-                    f"Path '{path}' does not have an allowed extension. "
-                    f"Allowed extensions: {', '.join(extensions)}"
+                    f"Path '{path}' matches exclude pattern "
+                    f"'{exclude_pattern}'."
                 )
 
-        # Patterns
-        if include_pattern:
-            if not fnmatch.fnmatch(path.name, include_pattern):
-                raise ValidationError(
-                    f"Path '{path}' does not match include pattern "
-                    f"'{include_pattern}'."
-                )
-        elif exclude_pattern and fnmatch.fnmatch(path.name, exclude_pattern):
-            raise ValidationError(
-                f"Path '{path}' matches exclude pattern "
-                f"'{exclude_pattern}'."
-            )
+            return path
 
-        return path
+        if is_single_value(value, (str, Path)):
+            return validate_path(value)
+        if is_tuple_value(value, (str, Path)):
+            return tuple(validate_path(v) for v in value)
+        if is_nested_tuple_value(value, (str, Path)):
+            return tuple(tuple(validate_path(v) for v in t) for t in value)
+        raise UnhandledValueError(value)
 
 
 def as_path(
@@ -139,6 +165,8 @@ def as_path(
 ) -> Decorator:
     """
     Transform a string to a valid path.
+
+    Supports: `str | Path`, `tuple[str | Path]` and `tuple[tuple[str | Path]]`
 
     Args:
         exists (bool, optional):