+++ click_extended/transform/to_uppercase.py
@@ -1,19 +1,37 @@
 """Transformation decorator to convert a string to uppercase."""
 
 from click_extended.core._child_node import ChildNode, ProcessContext
+from click_extended.errors import UnhandledValueError
 from click_extended.types import Decorator
+from click_extended.utils import (
+    is_nested_tuple_value,
+    is_single_value,
+    is_tuple_value,
+)
 
 
 class ToUppercase(ChildNode):
     """Transformation decorator to convert a string to uppercase."""
 
-    def process(self, value: str, context: ProcessContext) -> str:
-        return value.upper()
+    def process(
+        self,
+        value: str | tuple[str, ...] | tuple[tuple[str, ...], ...],
+        context: ProcessContext,
+    ) -> str | tuple[str, ...] | tuple[tuple[str, ...], ...]:
+        if is_single_value(value, str):
+            return value.upper()
+        if is_tuple_value(value, str):
+            return tuple(v.upper() for v in value)
+        if is_nested_tuple_value(value, str):
+            return tuple(tuple(v.upper() for v in t) for t in value)
+        raise UnhandledValueError(value)
 
 
 def to_uppercase() -> Decorator:
     """
     Convert a value to uppercase.
+
+    Supports `str`, `tuple[str]` and `tuple[tuple[str]]`.
 
     Returns:
         Decorator: